# docs/API_SPEC.yaml
openapi: 3.0.3
info:
  title: MyApi - Personal Identity & Data Gateway
  version: 1.0.0
  description: |
    MyApi provides a secure, privacy-first API gateway to access user identity, preferences, 
    and connected data sources. It serves as a Single Source of Truth (SSOT) for AI entities 
    and applications communicating with the user.
    
    ## Architecture
    - **Vault**: Dynamic context storage (identity docs, connectors, preferences)
    - **Gateway**: API surface with access control and token management
    - **Personal Brain**: Logic layer that routes requests and enforces privacy
    
    ## Security
    - Zero-Knowledge architecture
    - Scope-based access control
    - Full audit logging
    - Token revocation capabilities
    
  contact:
    name: MyApi Support
  license:
    name: Proprietary
servers:
  - url: https://api.myapi.io/v1
    description: Production server
  - url: https://staging-api.myapi.io/v1
    description: Staging server
  - url: http://localhost:3000/v1
    description: Local development server
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Personal Token (full access) or Guest Token (scoped access).
        Include in Authorization header: `Bearer <token>`
  schemas:
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: "UNAUTHORIZED"
        message:
          type: string
          example: "Invalid or expired token"
        details:
          type: object
          additionalProperties: true
    Meta:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
          example: "2026-02-18T21:00:00Z"
        version:
          type: string
          example: "1.0.0"
        requestId:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
        rateLimit:
          type: object
          properties:
            limit:
              type: integer
              example: 1000
            remaining:
              type: integer
              example: 999
            reset:
              type: integer
              format: int64
              description: Unix timestamp when rate limit resets
              example: 1708293600
...
    Token:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "b2c3d4e5-f6a7-4890-b123-c4d5e6f7a890"
        type:
          type: string
          enum: [personal, guest]
          example: "guest"
        name:
          type: string
          description: Human-readable name for the token
          example: "Calendar Integration"
        scope:
          type: array
          items:
            type: string
          example: ["identity:professional", "availability:read"]
          description: |
            Scopes define access permissions:
            - identity:read - Basic identity info
            - identity:professional - Professional profile
            - identity:personal - Personal info
            - availability:read - Calendar/availability
            - preferences:read - User preferences
            - connectors:read - List connectors
            - connectors:write - Add/remove connectors
            - audit:read - View audit logs
            - tokens:read - List tokens
            - tokens:write - Create/revoke tokens
        createdAt:
          type: string
          format: date-time
        expiresAt:
          type: string
          format: date-time
          nullable: true
        lastUsed:
          type: string
          format: date-time
          nullable: true
        usageCount:
          type: integer
          example: 42
    TokenCreateRequest:
      type: object
      required:
        - name
        - scope
      properties:
        name:
          type: string
          example: "Third-party calendar app"
        scope:
          type: array
          items:
            type: string
          example: ["identity:professional", "availability:read"]
        expiresIn:
          type: integer
          description: Token lifetime in seconds (optional, null = no expiry)
          example: 2592000
          nullable: true
...
    Connector:
      type: object
      properties:
        id:
          type: string
          format: uuid
        type:
          type: string
          enum: [github, linkedin, calendar, journal, notion, gdrive, custom]
          example: "github"
        name:
          type: string
          example: "GitHub Profile"
        status:
          type: string
          enum: [active, inactive, error]
          example: "active"
        readOnly:
          type: boolean
          example: true
          description: Whether this connector is read-only
        createdAt:
          type: string
          format: date-time
        lastSync:
          type: string
          format: date-time
          nullable: true
        config:
          type: object
          description: Connector-specific configuration (sanitized)
          additionalProperties: true
    ConnectorCreateRequest:
      type: object
      required:
        - type
        - name
        - config
      properties:
        type:
          type: string
          enum: [github, linkedin, calendar, journal, notion, gdrive, custom]
        name:
          type: string
          example: "My GitHub"
        config:
          type: object
          description: Connector-specific configuration
          additionalProperties: true
          example:
            username: "octocat"
            apiToken: "ghp_xxxxxxxxxxxxx"
...
  responses:
    UnauthorizedError:
      description: Authentication token is missing or invalid
      headers:
        X-RateLimit-Limit:
          schema:
            type: integer
        X-RateLimit-Remaining:
          schema:
            type: integer
        X-RateLimit-Reset:
          schema:
            type: integer
      content:
        application/json:
          schema:
            type: object
            properties:
              errors:
                type: array
                items:
                  $ref: '#/components/schemas/Error'
              meta:
                $ref: '#/components/schemas/Meta'
    ForbiddenError:
      description: Insufficient permissions for this operation
      headers:
        X-RateLimit-Limit:
          schema:
            type: integer
        X-RateLimit-Remaining:
          schema:
            type: integer
        X-RateLimit-Reset:
          schema:
            type: integer
      content:
        application/json:
          schema:
            type: object
            properties:
              errors:
                type: array
                items:
                  $ref: '#/components/schemas/Error'
              meta:
                $ref: '#/components/schemas/Meta'
    RateLimitError:
      description: Rate limit exceeded
      headers:
        X-RateLimit-Limit:
          schema:
            type: integer
        X-RateLimit-Remaining:
          schema:
            type: integer
            example: 0
        X-RateLimit-Reset:
          schema:
            type: integer
        Retry-After:
          schema:
            type: integer
      content:
        application/json:
          schema:
            type: object
            properties:
              errors:
                type: array
                items:
                  $ref: '#/components/schemas/Error'
              meta:
                $ref: '#/components/schemas/Meta'
  parameters:
    TokenIdParam:
      name: id
      in: path
      required: true
      description: Token ID
      schema:
        type: string
        format: uuid
security:
  - BearerAuth: []
paths:
...
  /tokens:
    get:
      tags:
        - Tokens
      summary: List active tokens
      description: |
        Lists all active tokens. Only accessible with a personal token.
        Returns token metadata without exposing the actual token values.
      operationId: listTokens
      parameters:
        - name: type
          in: query
          description: Filter by token type
          schema:
            type: string
            enum: [personal, guest]
      responses:
        '200':
          description: List of tokens
          headers:
            X-RateLimit-Limit:
              schema:
                type: integer
            X-RateLimit-Remaining:
              schema:
                type: integer
            X-RateLimit-Reset:
              schema:
                type: integer
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Token'
                  meta:
                    $ref: '#/components/schemas/Meta'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          description: Only personal tokens can list all tokens
          content:
            application/json:
              schema:
                type: object
                properties:
                  errors:
                    type: array
                    items:
                      $ref: '#/components/schemas/Error'
                  meta:
                    $ref: '#/components/schemas/Meta'
        '429':
          $ref: '#/components/responses/RateLimitError'

    post:
      tags:
        - Tokens
      summary: Create a new guest token
      description: |
        Creates a new guest token with specified scope and expiration.
        Only accessible with a personal token.
      operationId: createToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TokenCreateRequest'
      responses:
        '201':
          description: Token created successfully
          headers:
            X-RateLimit-Limit:
              schema:
                type: integer
            X-RateLimit-Remaining:
              schema:
                type: integer
            X-RateLimit-Reset:
              schema:
                type: integer
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    allOf:
                      - $ref: '#/components/schemas/Token'
                      - type: object
                        properties:
                          token:
                            type: string
                            description: The actual token value (only returned on creation)
                            example: "myapi_guest_abc123def456..."
                  meta:
                    $ref: '#/components/schemas/Meta'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                type: object
                properties:
                  errors:
                    type: array
                    items:
                      $ref: '#/components/schemas/Error'
                  meta:
                    $ref: '#/components/schemas/Meta'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          description: Only personal tokens can create guest tokens
          content:
            application/json:
              schema:
                type: object
                properties:
                  errors:
                    type: array
                    items:
                      $ref: '#/components/schemas/Error'
                  meta:
                    $ref: '#/components/schemas/Meta'
        '429':
          $ref: '#/components/responses/RateLimitError'
  /tokens/{id}:
    delete:
      tags:
        - Tokens
      summary: Revoke a token
      description: |
        Revokes a token immediately. The token will no longer be valid for any requests.
        Only accessible with a personal token.
      operationId: revokeToken
      parameters:
        - $ref: '#/components/parameters/TokenIdParam'
      responses:
        '204':
          description: Token revoked successfully
          headers:
            X-RateLimit-Limit:
              schema:
                type: integer
            X-RateLimit-Remaining:
              schema:
                type: integer
            X-RateLimit-Reset:
              schema:
                type: integer
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          description: Only personal tokens can revoke tokens
          content:
            application/json:
              schema:
                type: object
                properties:
                  errors:
                    type: array
                    items:
                      $ref: '#/components/schemas/Error'
                  meta:
                    $ref: '#/components/schemas/Meta'
        '404':
          description: Token not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  errors:
                    type: array
                    items:
                      $ref: '#/components/schemas/Error'
                  meta:
                    $ref: '#/components/schemas/Meta'
        '429':
          $ref: '#/components/responses/RateLimitError'
...
  /connectors:
    get:
      tags:
        - Connectors
      summary: List active connectors
      description: Returns a list of all active data source connectors
      operationId: listConnectors
      parameters:
        - name: type
          in: query
          description: Filter by connector type
          schema:
            type: string
            enum: [github, linkedin, calendar, journal, notion, gdrive, custom]
        - name: status
          in: query
          description: Filter by status
          schema:
            type: string
            enum: [active, inactive, error]
      responses:
        '200':
          description: List of connectors
          headers:
            X-RateLimit-Limit:
              schema:
                type: integer
            X-RateLimit-Remaining:
              schema:
                type: integer
            X-RateLimit-Reset:
              schema:
                type: integer
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Connector'
                  meta:
                    $ref: '#/components/schemas/Meta'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '429':
          $ref: '#/components/responses/RateLimitError'

    post:
      tags:
        - Connectors
      summary: Add a connector
      description: |
        Adds a new data source connector.
        Requires connectors:write scope.
      operationId: createConnector
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConnectorCreateRequest'
      responses:
        '201':
          description: Connector created successfully
          headers:
            X-RateLimit-Limit:
              schema:
                type: integer
            X-RateLimit-Remaining:
              schema:
                type: integer
            X-RateLimit-Reset:
              schema:
                type: integer
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Connector'
                  meta:
                    $ref: '#/components/schemas/Meta'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                type: object
                properties:
                  errors:
                    type: array
                    items:
                      $ref: '#/components/schemas/Error'
                  meta:
                    $ref: '#/components/schemas/Meta'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          description: Requires connectors:write scope
          content:
            application/json:
              schema:
                type: object
                properties:
                  errors:
                    type: array
                    items:
                      $ref: '#/components/schemas/Error'
                  meta:
                    $ref: '#/components/schemas/Meta'
        '429':
          $ref: '#/components/responses/RateLimitError'